[V041] Minor optimizations
