+++ setup.py
@@ -2,11 +2,10 @@ from setuptools import setup, find_packages
 
 setup(
     name="shell_gpt",
-    version="0.4",
+    version="0.4.1",
     packages=find_packages(),
     py_modules=[
         "sgpt",
-        "utils"
     ],
     install_requires=[
         "typer~=0.7.0",
@@ -19,8 +18,8 @@ setup(
         ],
     },
     author="Farkhod Sadykov",
-    author_email="farkhod.sadykov@gmail.com",
-    description="CLI App allows to query OpenAI Chat GPT models using API.",
+    author_email="farkhod@sadykov.dev",
+    description="CLI App allows to query OpenAI GPT-3 models using API.",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
     url="https://github.com/ther1d/shell_gpt",
